LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.measure.fontBoundingBox-expected.txt

2d.text.measure.fontBoundingBox

Testing fontBoundingBox measurements


FAIL Testing fontBoundingBox measurements assert_approx_equals: unexpected fontBoundingBoxAscent expected 85.205078125 +/- 1 but got 38
